crawler/requirements.txt
backport-ipaddress==0.1
certifi==2018.1.18
chardet==3.0.4
dnet==1.12
dnspython==1.15.0
idna==2.6
MySQL-python==1.2.5
netaddr==0.7.19
pcapy==0.11.1
peewee==3.1.5
pyasn==1.6.0b1
requests==2.18.4
scapy==2.3.3
urllib3==1.22
